refactor: consolidate duplicated test fakes into conftest.py
