+++ backend/tests/api/test_coaching.py
@@ -11,92 +11,7 @@ from app.infrastructure.vlm.solution_coaching_client import CoachingVLMResult
 from app.infrastructure.storage.mongo import COACHING_CONVERSATIONS_COLLECTION
 from app.main import create_app
 from app.presentation.deps import get_current_user, get_database
-
-class FakeCursor:
-    def __init__(self, documents):
-        self._documents = [deepcopy(doc) for doc in documents]
-    async def to_list(self, length=None):
-        return [deepcopy(doc) for doc in self._documents]
-
-class FakeCollection:
-    def __init__(self):
-        self._documents = []
-
-    def seed(self, *documents):
-        self._documents.extend(deepcopy(list(documents)))
-
-    async def find_one(self, query, sort=None):
-        docs = []
-        for doc in self._documents:
-            if all(doc.get(k) == v for k, v in query.items() if k != "items.problemId"):
-                if "items.problemId" in query:
-                    val = query["items.problemId"]
-                    if not any(val == item.get("problemId") for item in doc.get("items", [])):
-                        continue
-                docs.append(deepcopy(doc))
-        if not docs:
-            return None
-        if sort:
-            docs.sort(key=lambda x: x.get(sort[0][0]), reverse=(sort[0][1] == -1))
-        return docs[0]
-
-    def find(self, query):
-        docs = [d for d in self._documents if all(d.get(k) == v for k, v in query.items())]
-        return FakeCursor(docs)
-
-    async def insert_one(self, document):
-        doc = deepcopy(document)
-        if "_id" not in doc:
-            doc["_id"] = ObjectId()
-        self._documents.append(doc)
-        class Ret:
-            inserted_id = doc["_id"]
-        return Ret()
-
-    async def update_one(self, query, update, upsert=False):
-        for doc in self._documents:
-            if all(doc.get(k) == v for k, v in query.items()):
-                for k, v in update.get("$set", {}).items():
-                    doc[k] = deepcopy(v)
-                return
-        if upsert:
-            new_doc = deepcopy(query)
-            for k, v in update.get("$set", {}).items():
-                new_doc[k] = deepcopy(v)
-            if "_id" not in new_doc:
-                new_doc["_id"] = ObjectId()
-            self._documents.append(new_doc)
-
-    async def delete_one(self, query):
-        self._documents = [d for d in self._documents if not all(d.get(k) == v for k, v in query.items())]
-
-    async def count_documents(self, query):
-        count = 0
-        for doc in self._documents:
-            match = True
-            for k, v in query.items():
-                if k == "items.problemId":
-                    if not any(v == item.get("problemId") for item in doc.get("items", [])):
-                        match = False
-                        break
-                elif doc.get(k) != v:
-                    match = False
-                    break
-            if match:
-                count += 1
-        return count
-
-class FakeDatabase:
-    def __init__(self):
-        self._collections = {
-            "problems": FakeCollection(),
-            "exams": FakeCollection(),
-            "canonical_solutions": FakeCollection(),
-            "practice_attempts": FakeCollection(),
-            "coaching_conversations": FakeCollection(),
-        }
-    def __getitem__(self, name):
-        return self._collections[name]
+from tests.api.conftest import FakeDatabase
 
 @pytest_asyncio.fixture
 async def coaching_app() -> FastAPI:
