+++ backend/tests/api/test_folders.py
@@ -14,138 +14,7 @@ from httpx import ASGITransport, AsyncClient
 from app.main import create_app
 from app.presentation.deps import get_current_user, get_database
 from app.presentation.errors import ApiError
-
-
-class FakeInsertOneResult:
-    def __init__(self, inserted_id: Any) -> None:
-        self.inserted_id = inserted_id
-
-
-class FakeUpdateResult:
-    def __init__(self, modified_count: int) -> None:
-        self.modified_count = modified_count
-
-
-class FakeDeleteResult:
-    def __init__(self, deleted_count: int) -> None:
-        self.deleted_count = deleted_count
-
-
-class FakeCursor:
-    def __init__(self, documents: list[dict[str, Any]]) -> None:
-        self._documents = [deepcopy(document) for document in documents]
-
-    def sort(self, field: str, direction: int) -> FakeCursor:
-        reverse = direction < 0
-        self._documents.sort(
-            key=lambda document: cast(Any, document.get(field, "")),
-            reverse=reverse,
-        )
-        return self
-
-    def limit(self, amount: int) -> FakeCursor:
-        return self
-
-    async def to_list(self, length: int | None = None) -> list[dict[str, Any]]:
-        return [deepcopy(document) for document in self._documents]
-
-
-class FakeCollection:
-    def __init__(self) -> None:
-        self._documents: list[dict[str, Any]] = []
-
-    def seed(self, *documents: dict[str, Any]) -> None:
-        self._documents.extend(deepcopy(list(documents)))
-
-    async def find_one(self, query: dict[str, Any]) -> dict[str, Any] | None:
-        for document in self._documents:
-            if _matches(document, query):
-                return deepcopy(document)
-        return None
-
-    async def insert_one(self, document: dict[str, Any]) -> FakeInsertOneResult:
-        stored = deepcopy(document)
-        if "_id" not in stored:
-            stored["_id"] = ObjectId()
-        self._documents.append(stored)
-        return FakeInsertOneResult(stored["_id"])
-
-    async def update_one(self, query: dict[str, Any], update: dict[str, Any]) -> FakeUpdateResult:
-        for document in self._documents:
-            if _matches(document, query):
-                for key, value in update.get("$set", {}).items():
-                    document[key] = deepcopy(value)
-                return FakeUpdateResult(1)
-        return FakeUpdateResult(0)
-
-    async def delete_one(self, query: dict[str, Any]) -> FakeDeleteResult:
-        for i, document in enumerate(self._documents):
-            if _matches(document, query):
-                del self._documents[i]
-                return FakeDeleteResult(1)
-        return FakeDeleteResult(0)
-
-    async def count_documents(self, query: dict[str, Any]) -> int:
-        return sum(1 for document in self._documents if _matches(document, query))
-
-    def find(self, query: dict[str, Any], projection: dict[str, Any] | None = None) -> FakeCursor:
-        matches = [document for document in self._documents if _matches(document, query)]
-        return FakeCursor(matches)
-
-
-class FakeDatabase:
-    def __init__(self) -> None:
-        self._collections = {
-            "problems": FakeCollection(),
-            "tags": FakeCollection(),
-            "folders": FakeCollection(),
-            "users": FakeCollection(),
-            "sessions": FakeCollection(),
-        }
-
-    def __getitem__(self, name: str) -> FakeCollection:
-        return self._collections[name]
-
-
-def _matches(document: dict[str, Any], query: dict[str, Any]) -> bool:
-    for key, value in query.items():
-        actual = document.get(key)
-        if isinstance(value, dict):
-            if "$in" in value:
-                if actual not in value["$in"]:
-                    return False
-                continue
-            if "$ne" in value:
-                if actual == value["$ne"]:
-                    return False
-                continue
-            if "$regex" in value:
-                import re
-                pattern = value["$regex"]
-                options = value.get("$options", "")
-                flags = re.IGNORECASE if "i" in options else 0
-                if not re.search(pattern, str(actual or ""), flags):
-                    return False
-                continue
-        if isinstance(actual, list):
-            if value not in actual:
-                return False
-            continue
-        if actual != value:
-            return False
-    return True
-
-
-def make_user(user_id: ObjectId, username: str) -> dict[str, Any]:
-    now = datetime.now(UTC)
-    return {
-        "_id": user_id,
-        "username": username,
-        "status": "active",
-        "createdAt": now,
-        "updatedAt": now,
-        "lastLoginAt": None,
-    }
+from tests.api.conftest import FakeDatabase, make_user
 
 
 def make_folder(
