+++ backend/tests/api/test_ingestion.py
@@ -26,150 +26,7 @@ from app.presentation.deps import (
     get_database,
 )
 from app.presentation import ingestion as ingestion_presentation
-
-
-class FakeInsertOneResult:
-    def __init__(self, inserted_id: Any) -> None:
-        self.inserted_id = inserted_id
-
-
-class FakeUpdateResult:
-    def __init__(self, modified_count: int) -> None:
-        self.modified_count = modified_count
-
-
-class FakeDeleteResult:
-    def __init__(self, deleted_count: int) -> None:
-        self.deleted_count = deleted_count
-
-
-class FakeCursor:
-    def __init__(self, documents: list[dict[str, Any]]) -> None:
-        self._documents = [deepcopy(document) for document in documents]
-
-    def sort(self, field: str, direction: int) -> FakeCursor:
-        reverse = direction < 0
-        self._documents.sort(key=lambda document: document.get(field), reverse=reverse)
-        return self
-
-    async def to_list(self, length: int | None = None) -> list[dict[str, Any]]:
-        if length is None:
-            return [deepcopy(document) for document in self._documents]
-        return [deepcopy(document) for document in self._documents[:length]]
-
-
-def _matches(document: dict[str, Any], query: dict[str, Any]) -> bool:
-    for key, value in query.items():
-        actual = document.get(key)
-        if isinstance(value, dict):
-            allowed_values = value.get("$in")
-            if allowed_values is not None:
-                if actual not in allowed_values:
-                    return False
-                continue
-            return False
-        if isinstance(actual, list):
-            if value not in actual:
-                return False
-            continue
-        if actual != value:
-            return False
-    return True
-
-
-def _set_nested(document: dict[str, Any], path: str, value: Any) -> None:
-    parts = path.split(".")
-    target = document
-    for part in parts[:-1]:
-        if part not in target:
-            target[part] = {}
-        target = target[part]
-    target[parts[-1]] = value
-
-
-class FakeCollection:
-    def __init__(self) -> None:
-        self._documents: list[dict[str, Any]] = []
-        self._insert_one_error: Exception | None = None
-        self._delete_one_error: Exception | None = None
-
-    def seed(self, *documents: dict[str, Any]) -> None:
-        self._documents.extend(deepcopy(list(documents)))
-
-    async def find_one(self, query: dict[str, Any]) -> dict[str, Any] | None:
-        for document in self._documents:
-            if _matches(document, query):
-                return deepcopy(document)
-        return None
-
-    async def insert_one(self, document: dict[str, Any]) -> FakeInsertOneResult:
-        if self._insert_one_error is not None:
-            raise self._insert_one_error
-        stored_document = deepcopy(document)
-        if "_id" not in stored_document:
-            stored_document["_id"] = ObjectId()
-        self._documents.append(stored_document)
-        return FakeInsertOneResult(stored_document["_id"])
-
-    async def insert_many(self, documents: list[dict[str, Any]], ordered: bool = True) -> None:
-        for document in documents:
-            stored_document = deepcopy(document)
-            if "_id" not in stored_document:
-                stored_document["_id"] = ObjectId()
-            self._documents.append(stored_document)
-
-    async def update_one(self, query: dict[str, Any], update: dict[str, Any]) -> FakeUpdateResult:
-        for document in self._documents:
-            if _matches(document, query):
-                for key, value in update.get("$set", {}).items():
-                    _set_nested(document, key, deepcopy(value))
-                return FakeUpdateResult(1)
-        return FakeUpdateResult(0)
-
-    async def find_one_and_update(
-        self,
-        query: dict[str, Any],
-        update: dict[str, Any],
-    ) -> dict[str, Any] | None:
-        for document in self._documents:
-            if _matches(document, query):
-                original_document = deepcopy(document)
-                for key, value in update.get("$set", {}).items():
-                    _set_nested(document, key, deepcopy(value))
-                return original_document
-        return None
-
-    async def delete_one(self, query: dict[str, Any]) -> FakeDeleteResult:
-        if self._delete_one_error is not None:
-            raise self._delete_one_error
-        for index, document in enumerate(self._documents):
-            if _matches(document, query):
-                del self._documents[index]
-                return FakeDeleteResult(1)
-        return FakeDeleteResult(0)
-
-    def find(self, query: dict[str, Any]) -> FakeCursor:
-        matching = [document for document in self._documents if _matches(document, query)]
-        return FakeCursor(matching)
-
-    async def count_documents(self, query: dict[str, Any]) -> int:
-        return sum(1 for document in self._documents if _matches(document, query))
-
-
-class FakeDatabase:
-    def __init__(self) -> None:
-        self._collections = {
-            "users": FakeCollection(),
-            "sessions": FakeCollection(),
-            "ingestion_previews": FakeCollection(),
-            "problems": FakeCollection(),
-            "tags": FakeCollection(),
-            "solution_generation_tasks": FakeCollection(),
-            "canonical_solutions": FakeCollection(),
-        }
-
-    def __getitem__(self, name: str) -> FakeCollection:
-        return self._collections[name]
+from tests.api.conftest import FakeDatabase
 
 
 class FakeStorage:
