+++ backend/tests/api/test_problems.py
@@ -15,132 +15,7 @@ from app.infrastructure.storage.s3 import StorageObjectNotFoundError
 from app.main import create_app
 from app.presentation.deps import get_current_user, get_database, get_s3_storage
 from app.presentation.errors import ApiError
-
-
-class FakeInsertOneResult:
-    def __init__(self, inserted_id: Any) -> None:
-        self.inserted_id = inserted_id
-
-
-class FakeUpdateResult:
-    def __init__(self, modified_count: int) -> None:
-        self.modified_count = modified_count
-
-
-class FakeCursor:
-    def __init__(self, documents: list[dict[str, Any]]) -> None:
-        self._documents = [deepcopy(document) for document in documents]
-        self._skip = 0
-        self._limit: int | None = None
-
-    def sort(self, field: str | list[tuple[str, int]], direction: int | None = None) -> FakeCursor:
-        sort_fields = field if isinstance(field, list) else [(field, direction or 1)]
-        for sort_field, sort_direction in reversed(sort_fields):
-            reverse = sort_direction < 0
-            self._documents.sort(
-                key=lambda document: cast(Any, document.get(sort_field)),
-                reverse=reverse,
-            )
-        return self
-
-    def skip(self, amount: int) -> FakeCursor:
-        self._skip = amount
-        return self
-
-    def limit(self, amount: int) -> FakeCursor:
-        self._limit = amount
-        return self
-
-    async def to_list(self, length: int | None = None) -> list[dict[str, Any]]:
-        documents = self._documents[self._skip :]
-        effective_limit = self._limit if self._limit is not None else length
-        if effective_limit is not None:
-            documents = documents[:effective_limit]
-        return [deepcopy(document) for document in documents]
-
-
-class FakeCollection:
-    def __init__(self) -> None:
-        self._documents: list[dict[str, Any]] = []
-
-    def seed(self, *documents: dict[str, Any]) -> None:
-        self._documents.extend(deepcopy(list(documents)))
-
-    async def find_one(self, query: dict[str, Any]) -> dict[str, Any] | None:
-        for document in self._documents:
-            if _matches(document, query):
-                return deepcopy(document)
-        return None
-
-    async def insert_one(self, document: dict[str, Any]) -> FakeInsertOneResult:
-        stored = deepcopy(document)
-        if "_id" not in stored:
-            stored["_id"] = ObjectId()
-        self._documents.append(stored)
-        return FakeInsertOneResult(stored["_id"])
-
-    async def insert_many(self, documents: list[dict[str, Any]], ordered: bool = True) -> None:
-        for document in documents:
-            stored = deepcopy(document)
-            if "_id" not in stored:
-                stored["_id"] = ObjectId()
-            self._documents.append(stored)
-
-    async def update_one(self, query: dict[str, Any], update: dict[str, Any]) -> FakeUpdateResult:
-        for document in self._documents:
-            if _matches(document, query):
-                for key, value in update.get("$set", {}).items():
-                    document[key] = deepcopy(value)
-                return FakeUpdateResult(1)
-        return FakeUpdateResult(0)
-
-    async def update_many(self, query: dict[str, Any], update: dict[str, Any]) -> FakeUpdateResult:
-        count = 0
-        for document in self._documents:
-            if _matches(document, query):
-                for key, value in update.get("$set", {}).items():
-                    document[key] = deepcopy(value)
-                count += 1
-        return FakeUpdateResult(count)
-
-    async def count_documents(self, query: dict[str, Any]) -> int:
-        return sum(1 for document in self._documents if _matches(document, query))
-
-    def find(self, query: dict[str, Any], projection: dict[str, Any] | None = None) -> FakeCursor:
-        matches = [document for document in self._documents if _matches(document, query)]
-        return FakeCursor(matches)
-
-    async def distinct(self, field: str, query: dict[str, Any]) -> list[Any]:
-        values: list[Any] = []
-        seen: set[Any] = set()
-        for document in self._documents:
-            if not _matches(document, query):
-                continue
-            current = document.get(field, [])
-            iterable = current if isinstance(current, list) else [current]
-            for value in iterable:
-                if value in seen:
-                    continue
-                seen.add(value)
-                values.append(value)
-        return values
-
-
-class FakeDatabase:
-    def __init__(self) -> None:
-        self._collections = {
-            "problems": FakeCollection(),
-            "tags": FakeCollection(),
-            "folders": FakeCollection(),
-            "ingestion_previews": FakeCollection(),
-            "users": FakeCollection(),
-            "sessions": FakeCollection(),
-            "solution_generation_tasks": FakeCollection(),
-            "canonical_solutions": FakeCollection(),
-        }
-
-    def __getitem__(self, name: str) -> FakeCollection:
-        return self._collections[name]
+from tests.api.conftest import FakeDatabase, make_user
 
 
 class FakeStorage:
@@ -155,54 +30,6 @@ class FakeStorage:
         if payload is None:
             raise StorageObjectNotFoundError(key)
         return payload
-
-
-def _matches(document: dict[str, Any], query: dict[str, Any]) -> bool:
-    for key, value in query.items():
-        if key == "$or":
-            if not any(_matches(document, sub) for sub in value):
-                return False
-            continue
-        actual = document.get(key)
-        if isinstance(value, dict):
-            if "$in" in value:
-                if actual not in value["$in"]:
-                    return False
-                continue
-            if "$regex" in value:
-                import re
-
-                pattern = value["$regex"]
-                options = value.get("$options", "")
-                flags = 0
-                if "i" in options:
-                    flags |= re.IGNORECASE
-                if isinstance(actual, list):
-                    if not any(re.search(pattern, str(item), flags) for item in actual):
-                        return False
-                else:
-                    if not re.search(pattern, str(actual or ""), flags):
-                        return False
-                continue
-        if isinstance(actual, list):
-            if value not in actual:
-                return False
-            continue
-        if actual != value:
-            return False
-    return True
-
-
-def make_user(user_id: ObjectId, username: str) -> dict[str, Any]:
-    now = datetime.now(UTC)
-    return {
-        "_id": user_id,
-        "username": username,
-        "status": "active",
-        "createdAt": now,
-        "updatedAt": now,
-        "lastLoginAt": None,
-    }
 
 
 def make_preview(user_id: ObjectId, *, status: str = "ready") -> dict[str, Any]:
